requirements.txt
bs4
websockets
aiohttp
pyfiglet
asyncio
pyjokes
datetime
colorama
pillow
qrcode
youtube_dl
beautifulsoup4
requests
wget
py-cord
flask
discord.py-self